Move gnome-keyring credential helper from git-core to git
We can avoid pulling in libgnome-keyring for git-core installs.
Presumably, those who want a minimal git install won't miss a
credentials helper which is only useful with a graphical desktop.

It is also worth noting that libgnome-keyring is deprecated in favor of
libsecret.  A credential helper based on libsecret is included in git
2.11.0 but is not shipped in our packages (yet).