Split credential-gnome-keyring into separate rpm
Since the libsecret replaces the libgnome-keyring,
the libgnome-keyring is not necessary to keep it in the git package.
In addition, gnome-keyring is deprecated, so maybe sooner or later
will be removed from our packages too. For now, it could be split
into the separate rpm, so gnome-keyring will not be required
by the git rpm and users, who don't use gnome, will not have to
install it as dependency.

The git-all rpm includes requirement on git-gnome-keyring package.