fix GIT_SKIP_TESTS for EL8 s390x
The GIT_SKIP_TESTS variable does not support brace expansion.  It was my
mistake thinking that it did.  List the tests to skip properly.

If we had a longer list and *really* wanted to use brace expansion, we
could do something like this:

    GIT_SKIP_TESTS="$GIT_SKIP_TESTS $(echo t5300.{10,12,14} t5303.{5,7,11} t6300.{35,91,92})"

In this case, that's more characters _and_ more complexity, so it makes
no sense to use it.  (Even if it were shorter, it doesn't necessarily
justify the extra complexity.)

Expand the list of tests to skip to cover those which fail due to the
earlier skipped tests.

Additionally, GIT_SKIP_TESTS is (unintentionally) set on systems other
than EL8.  Fix the conditional to only skip these tests on s390x on EL8.