lint: update filter rules
We removed '%{_emacs_version}' in 3395646 (remove --with/--without emacs
build conditional, 2022-06-13).  Drop the unnecessary filter from the
rpmlint config.

Add filters for several new checks in rpmlint 2.x: files-duplicate;
package-with-huge-docs; and potential-bashisms.

Also ignore unused-direct-shlib-dependency for libpcre2.  While this
is accurate, the additional linking would be tricky to remove from the
upstream Makefile.  It would almost certainly not be worth the effort.

Lastly (even though it's the first line in the file), drop the unneeded
'from Config import *' directive.  The rpmlint config is no longer
loaded directly as python code (yay!).