re-enable t5559-http-fetch-smart-http2, BuildRequires: mod_ssl
The t5559-http-fetch-smart-http2 test was disabled in 7d21254 (update to
2.39.0-rc0, 2022-11-23), due to its flakiness.

During the 2.40 cycle, 8f2146dbf1 (t5559: make SSL/TLS the default,
2023-02-23) was added by Jeff King.  This aimed to make the test both
more robust and indicative of the real world.

Testing suggests it is a success.  Spurious failures of the test have
not been seen in a number of runs across supported Fedora and EL
releases.