replace pod2man file dep with perl-podlators package dep
The file dependency was added in a91c5fd (Skip test BuildRequires when
--without tests is used, 2018-10-11) for EL6 support (where pod2man was
in the perl package).  EL6 has been dead for years now, revert to the
package dependency.

In Fedora, 2600353 (Fix file dependencies, 2023-07-06) replaced
%{_bindir} with /usr/bin in order to comply with the guidelines and
support flatpak perl installs.  But this misses the point that the file
dependency is no longer needed.